+++ packages/ZbayLite/src/renderer/containers/widgets/channels/DirectMessagesHeader.js
@@ -7,7 +7,8 @@ import contactsSelectors from '../../../store/selectors/contacts'
 export const mapStateToProps = (state, props) => {
   const contact = contactsSelectors.contact(props.contactId)(state)
   return {
-    channel: Immutable.fromJS({ name: contact.get('username') })
+    channel: Immutable.fromJS({ name: contact.get('username') }),
+    directMessage: true
   }
 }