+++ packages/ZbayLite/src/renderer/store/handlers/directMessagesQueue.js
@@ -114,6 +114,9 @@ const _sendPendingDirectMessages = async (dispatch, getState) => {
         ))
         return
       }
+      if (recipientAddress.length === 35) {
+        return
+      }
       dispatch(removeMessage(key))
       await dispatch(operationsHandlers.epics.observeOperation({
         opId,